GPU: tgv3d.cu seeded the FP64 build with float trig
Audit of the eight GPU drivers that had never been checked against the
parent's corresponding case -- alfven, orszag_tang, rayleigh_benard, tgv3d,
tgv3d_bench, channel, box, bench -- after droplet.cu was found seeding an
interface twice too wide (1e50657). One finding.

src/tgv3d.cu built its initial condition with cosf/sinf unconditionally, so
an -DLBM_DOUBLE build seeded the entire velocity field to FLOAT accuracy and
capped what the FP64 run could resolve before the first step. src/
tgv3d_bench.cu gets this right and its comment says why, naming this file:

  "cos/sin rather than cosf/sinf: this runs once per cell at start-up, so the
   double-precision promotion in an FP32 build costs nothing measurable and
   the same source is correct when built with -DLBM_DOUBLE."

MEASURED, and it does not move a published number. FP64, 24^3, t* to 2, 2400
steps: every reported row of E/E0 and Psi/Psi0 is identical before and after,
so the effect is below the printed resolution -- under about 3e-6 relative on
E/E0, not zero. Normalised ratios cancel most of it and a 1e-7 perturbation
on a smooth field damps fast. The FP32 build is bit-identical.

WHAT THE AUDIT CLEARED, each against the parent's own seeding and parameter
derivation rather than against the GPU host reference:

  alfven          uy = -(b/sqrt(rho)) sin kx, b_1 = b sin kx, b_0 = B0,
                  damping (nu+eta)k^2/2                   -- matches Mhd.hpp
  orszag_tang     v0 = Ma/(2 sqrt2 sqrt3), b0 = 0.8 v0 (parent applies the
                  0.8 in the seed, here in b0 -- same field), nu = v0 M/Re,
                  dt = v0 (2 pi/M)                        -- matches
  rayleigh_benard parent picks u_c and derives nu, this picks nu and derives
                  g; both give Ra = g beta dT H^3/(nu D) exactly, ny = H+2
  tgv3d           IC, diagnostics and vorticity stencil identical -- the
                  cosf finding above is the only difference
  tgv3d_bench     dr = 3 u0^2/16, same IC and same E/Z scalings
  channel         G = 8 rho0 nu umax/H^2, rest IC, ny = H+2 -- matches
                  poiseuille.cpp
  box, bench      throughput benchmarks with no parent physical counterpart;
                  their correctness claim is the mass-drift column, not a
                  reference result, so there is no seeding to cross-check

Host tests pass in both precisions, all four suites.

+++ GPU/src/tgv3d.cu
@@ -41,9 +41,14 @@ struct TaylorGreenInit {
     const Real X = k * Real(x), Y = k * Real(y), Z = k * Real(z);
     Macro m;
     m.rho = Real(1);
-    m.ux  =  u0 * cosf(X) * sinf(Y) * sinf(Z);
-    m.uy  = -Real(0.5) * u0 * sinf(X) * cosf(Y) * sinf(Z);
-    m.uz  = -Real(0.5) * u0 * sinf(X) * sinf(Y) * cosf(Z);
+    // cos/sin, NOT cosf/sinf. This runs once per cell at start-up, so the
+    // double-precision promotion in an FP32 build costs nothing measurable,
+    // while cosf in an -DLBM_DOUBLE build would seed the whole velocity field
+    // to FLOAT accuracy and quietly cap what the FP64 run can resolve.
+    // src/tgv3d_bench.cu had this right and named this file; this file did not.
+    m.ux  =  u0 * cos(X) * sin(Y) * sin(Z);
+    m.uy  = -Real(0.5) * u0 * sin(X) * cos(Y) * sin(Z);
+    m.uz  = -Real(0.5) * u0 * sin(X) * sin(Y) * cos(Z);
     return m;
   }
 };
